# Optimized Yolov5/v8 Segmentation code