Refactored Yolov5/v8 Segmentation code to speed up post processing and mask rendering.

* Added generic buffer‑pooling via a segBufferPool map of named pools to reuse []uint8 buffers of arbitrary lengths.

* Eliminated all per‑frame make([]uint8,…) calls for intermediate arrays (matmulOut, segMask, allMask, cropBuf) by reusing pooled buffers, collapsing GC pauses and heap churn.

* Optimized matmul to write directly into a caller‑provided slice (from the pool), with both a straight‐line and a stride‐based parallel version—no allocations inside the hot loops.

* ROI‑based merging of per‑object masks at the model‑input resolution:

* Resize each object’s prototype mask once to the full model dims,

* Only walk the pixels inside that object’s bounding‐box ROI to merge its ID into a single allMask array,

* Replacing costly global segMask + cropMaskWithIDUint8 passes.

* One‑time segReverse at the end to map the combined model‑input mask back to the original image size, preserving padding removal and artifact filtering exactly as before.

The above has resulted in the following performance enhancement.

| Model|	Average Total Time	|Inference	|Post Processing	|Rendering|
| --- | --- | --- | --- | --- |
| YOLOv5s-seg - BEFORE | 207.5ms            | 57.0ms    | 107.6ms         | 75.9ms    |
 YOLOv5s-seg - AFTER | 106.1ms            | 45.0ms    | 68.4ms          | 4.8ms     |
| YOLOv8s-seg  - BEFORE | 232.5ms            | 67.9ms    | 113.3ms         | 73.4ms    |
| YOLOv8s-seg - AFTER | 122.6ms            | 52.7ms    | 69.3ms          | 4.3ms     |